update integration test for mongodb
