+++ .travis.yml
@@ -9,9 +9,9 @@ script:
 - echo "Running Unit Tests"
 - docker run -e APP_KEY -e APP_TOKEN -e board_id -e list_id_Done -e list_id_Pending -e list_id_ToDo todo-app:test Test/view_model_test.py
 
-# Docker Run to Execute Unit Integration Package - Removed as part of module 9 to migrate to MongoDB
-# - echo "Running Integration Tests"
-# - docker run -e APP_KEY -e APP_TOKEN -e board_id -e list_id_Done -e list_id_Pending -e list_id_ToDo todo-app:test Test/client_test.py
+# Docker Run to Execute Unit Integration Package
+- echo "Running Integration Tests"
+- docker run -e APP_KEY -e APP_TOKEN -e board_id -e list_id_Done -e list_id_Pending -e list_id_ToDo todo-app:test Test/client_test.py
 
 # Docker Run to Execute Unit E2E Selenium Package - Removed as part of module 9 to migrate to MongoDB
 # - echo "Running E2E Tests"
